# documentation/openapi.yaml
openapi: 3.0.3
info:
  title: Club Accounting API
  version: 1.0.0
  description: REST API for managing projects, accounts, and bookings.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /project:
    get:
      tags:
        - Project
      summary: List projects
      operationId: listProjects
      responses:
        '200':
          description: A list of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags:
        - Project
      summary: Create a project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /project/{projectName}:
    get:
      tags:
        - Project
      summary: Get a project by name
      operationId: getProject
      parameters:
        - name: projectName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
  /account:
    get:
      tags:
        - Account
      summary: List accounts for a project
      operationId: listAccounts
      parameters:
        - name: projectName
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      tags:
        - Account
      summary: Create an account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
  /account/{id}:
    get:
      tags:
        - Account
      summary: Get an account by id
      operationId: getAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
  /booking:
    get:
      tags:
        - Booking
      summary: List bookings for a project
      operationId: listBookings
      parameters:
        - name: projectName
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
    post:
      tags:
        - Booking
      summary: Create a booking
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
  /booking/{id}:
    get:
      tags:
        - Booking
      summary: Get a booking by id
      operationId: getBooking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Booking found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found
components:
  schemas:
    Project:
      type: object
      properties:
        projectName:
          type: string
        passwordHash:
          type: string
      required:
        - projectName
        - passwordHash
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountNumber:
          type: string
        name:
          type: string
        projectName:
          type: string
      required:
        - accountNumber
        - name
        - projectName
    Booking:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookingNumber:
          type: string
        date:
          type: string
          format: date
        text:
          type: string
        debitAccountId:
          type: integer
          format: int64
        creditAccountId:
          type: integer
          format: int64
        amount:
          type: number
          format: double
        projectName:
          type: string
      required:
        - bookingNumber
        - date
        - text
        - debitAccountId
        - creditAccountId
        - amount
        - projectName